First draft of sc intensities function
