--- src/Intensities/Types.jl
@@ -1,2 +1,0 @@
-abstract type IntensityFormula end
-
