+++ src/SampledCorrelations/Interpolation.jl
@@ -1,0 +1,98 @@
+abstract type InterpolationScheme{NumInterp} end
+struct NoInterp <: InterpolationScheme{1} end
+struct LinearInterp <: InterpolationScheme{8} end
+
+
+function interpolated_intensity(::SampledCorrelations, _, stencil_intensities, ::NoInterp) 
+    return only(stencil_intensities)
+end
+
+# N.B.: This interpolation is only valid when all three entries of m_target are in [0,1]
+function interpolated_intensity(::SampledCorrelations, m_target, stencil_intensities, ::LinearInterp) 
+    c000, c100, c010, c110, c001, c101, c011, c111 = stencil_intensities
+
+    xd, yd, zd = m_target
+
+    c00 = c000*(1-xd) + c100*xd
+    c01 = c001*(1-xd) + c101*xd
+    c10 = c010*(1-xd) + c110*xd
+    c11 = c011*(1-xd) + c111*xd
+
+    c0 = c00*(1-yd) + c10*yd
+    c1 = c01*(1-yd) + c11*yd
+
+    return c0*(1-zd) + c1*zd
+end
+
+function stencil_points(sc::SampledCorrelations, q, ::NoInterp)
+    Ls = size(sc.samplebuf)[2:4] 
+    m = round.(Int, Ls .* q)
+    im = map(i -> mod(m[i], Ls[i])+1, (1, 2, 3)) |> CartesianIndex{3}
+
+    return (m,), (im,)
+end
+
+
+function stencil_points(sc::SampledCorrelations, q, ::LinearInterp)
+    Ls = size(sc.samplebuf)[2:4] 
+    base = map(x -> floor(Int64, x), Ls .* q) 
+    offsets = (
+        (0, 0, 0),
+        (1, 0, 0),
+        (0, 1, 0), 
+        (1, 1, 0), 
+        (0, 0, 1),
+        (1, 0, 1), 
+        (0, 1, 1),
+        (1, 1, 1)
+    )
+    ms = map(x -> x .+ base, offsets) 
+    ims = map(ms) do m
+        map(i -> mod(m[i], Ls[i])+1, (1, 2, 3)) |> CartesianIndex{3}
+    end
+    return ms, ims
+end
+
+# Note that requests for intensities often come in lists of nearby q values.
+# Since the data is inherently discretized, this often results in repeated calls
+# for values at the same discrete points. Since basis reduction is done for each
+# of this calls, this results in a large amount of repeated calculation. This
+# function analyzes repetitions in advance and prunes them out. This is
+# ugly, but the speedup when tested on a few simple, realistic examples was
+# 3-5x.
+function pruned_stencil_info(sc::SampledCorrelations, qs, interp::InterpolationScheme{N}) where N
+    # Count the number of contiguous regions with unchanging values. If all
+    # values are unique, returns the length of q_info. Note comparison is on m
+    # values rather than index values and the m values are the first element of
+    # the a tuple, that is, we're checking x[1] == y[1] in the map.
+    m_info = map(q -> stencil_points(sc, q, interp), qs)
+    numregions = sum(map((x,y) -> x[1] == y[1] ? 0 : 1, m_info[1:end-1], m_info[2:end])) + 1
+    
+    # Remove repeated stencil points and count number of instances of each
+    ms_ref, idcs_ref = stencil_points(sc, qs[1], interp)
+    ms_all  = fill(ntuple(x->zero(Vec3), N), numregions)
+    ms_all[1] = ms_ref 
+    idcs_all = fill(ntuple(x->CartesianIndex((-1,-1,-1)), N), numregions)
+    idcs_all[1] = idcs_ref 
+    counts = zeros(Int64, numregions)
+    c = counts[1] = 1
+    for q in qs[2:end] 
+        ms, idcs = stencil_points(sc, q, interp)
+        if ms != ms_ref
+            ms_ref = ms 
+            c += 1
+            ms_all[c] =  ms
+            idcs_all[c] = idcs 
+        end
+        counts[c] += 1
+    end
+    @assert sum(counts) == length(m_info)
+
+    # Calculate corresponding wave vectors in absolute units
+    ks_all = map(ms_all) do ms
+        map(m -> sc.crystal.recipvecs * (m ./ sc.latsize), ms)
+    end
+
+    return (; ks_all, idcs_all, counts)
+end
+
